refactor: centralize EVPN runtime plan shape policy
