+++ ARCHITECTURE.md
@@ -56,7 +56,7 @@ unicast Linux FIB, the BFD socket actor, and the EVPN dataplane glue).
 - `bfd` is a pure sans-IO crate with zero internal deps: RFC 5880 control-packet codec plus the session state machine. Like `fsm`, it never imports tokio or touches a socket — the daemon binary's `src/bfd_runtime.rs` owns the UDP sockets, per-session timers, and discriminator demux (ADR-0067).
 - `transport` is the only crate that owns BGP peer TCP session I/O and drives the FSM. Other crates (`api`, `bmp`, `rpki`, `mrt`) run their own async tasks for gRPC serving, collector connections, RTR sessions, and dump I/O respectively.
 - `rib` and `policy` are independent of transport and fsm — they consume route update events.
-- `evpn` is the local-VTEP domain crate (ADR-0052, ADR-0055, ADR-0058). It depends only on `wire`. It does **not** depend on `rib` or `transport`, and it never programs the kernel — kernel reconciliation lives in `crates/evpn-linux` (ADR-0054, shipped Gate 7b/7b+1; Gate 9 IP-VRF readiness probe + Linux netlink dumps + `probe_ip_vrfs` trait surface). The bidirectional VTEP loop is wired in the daemon binary by `src/evpn_dataplane.rs` (downward: RIB best-path → kernel FDB; also publishes the `IpVrfTable` through `DataplaneIntent` so the reconciler can probe IRB readiness every pass), `src/evpn_originator/` (upward: kernel local-MAC / MAC+IP observations → BGP Type 2 originations), and `src/evpn_imet.rs` (Type 3 IMET originations). RR-only deployments (empty `[[evpn_instances]]` and empty `[[evpn_ip_vrfs]]`) spawn no background tasks for either direction.
+- `evpn` is the local-VTEP domain crate (ADR-0052, ADR-0055, ADR-0058). It depends only on `wire`. It does **not** depend on `rib` or `transport`, and it never programs the kernel — kernel reconciliation lives in `crates/evpn-linux` (ADR-0054, shipped Gate 7b/7b+1; Gate 9 IP-VRF readiness probe + Linux netlink dumps + `probe_ip_vrfs` trait surface). The crate also owns the pure supported-plan-shape classifier used by both the daemon converger and plan decomposer; actor execution remains in the binary. This boundary removes the old `bench-internals` library-target mirror of ten EVPN actor modules without changing the default daemon build. The bidirectional VTEP loop is wired in the daemon binary by `src/evpn_dataplane.rs` (downward: RIB best-path → kernel FDB; also publishes the `IpVrfTable` through `DataplaneIntent` so the reconciler can probe IRB readiness every pass), `src/evpn_originator/` (upward: kernel local-MAC / MAC+IP observations → BGP Type 2 originations), and `src/evpn_imet.rs` (Type 3 IMET originations). RR-only deployments (empty `[[evpn_instances]]` and empty `[[evpn_ip_vrfs]]`) spawn no background tasks for either direction.
 - `api` provides the gRPC server; the binary crate (`src/main.rs`) wires everything together.
 
 ---
