+++ tests/lib_module_mirror.rs
@@ -25,6 +25,16 @@ const NOT_MIRRORED: &[&str] = &[
     "config_persister",
     "config_transaction_control",
     "confirm_journal",
+    "evpn_ack",
+    "evpn_dataplane",
+    "evpn_es_drain",
+    "evpn_es_link_drain",
+    "evpn_imet",
+    "evpn_l3_originator",
+    "evpn_originator",
+    "evpn_runtime_converger",
+    "evpn_segment",
+    "evpn_svi",
     "fib_runtime",
     "fib_table_control",
     "gnmi_set_bridge",
